(api-cat) Rename restapi package to controller
