+++ applications/api-cat/src/main/java/no/acat/controller/ApiRestController.java
@@ -1,4 +1,4 @@
-package no.acat.restapi;
+package no.acat.controller;
 
 import io.swagger.annotations.ApiOperation;
 import no.acat.model.ApiDocument;
