+++ applications/api-cat/src/main/java/no/acat/controller/ControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package no.acat.restapi;
+package no.acat.controller;
 
 import no.dcat.webutils.exceptions.GlobalControllerExceptionHandler;
 import org.springframework.web.bind.annotation.ControllerAdvice;
