+++ applications/api-cat/src/test/java/no/acat/controller/ConvertControllerTest.java
@@ -1,4 +1,4 @@
-package no.acat.restapi;
+package no.acat.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import no.acat.service.ParserService;
